requirements.txt
tensorflow==2.20.0
numpy==1.26.4
pandas==2.2.2
opencv-python==4.10.0.84
Pillow==10.4.0
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.1
tqdm==4.66.5
tf-keras-vis==0.8.6
streamlit==1.39.0
kaggle==1.6.12
